changed all occurances of "unsigned long" into uint32_t
    - makes sure integers have save width across platforms
    - addresses issue eudaq/eudaq#61

+++ main/include/eudaq/Event.hh
@@ -104,8 +104,8 @@ namespace eudaq {
       }
 
       typedef Event * (* event_creator)(Deserializer & ds);
-      static void Register(unsigned long id, event_creator func);
-      static event_creator GetCreator(unsigned long id);
+      static void Register(uint32_t id, event_creator func);
+      static event_creator GetCreator(uint32_t id);
 
     private:
       typedef std::map<unsigned long, event_creator> map_t;
