+++ producers/altro/src/AltroProducer.cc
@@ -299,7 +299,7 @@ void AltroProducer::SetRunActive(bool activestatus)
     pthread_mutex_unlock( &m_runactive_mutex );    
 }
 
-void AltroProducer::Event(volatile unsigned long *altrodata, int length)
+void AltroProducer::Event(volatile uint32_t *altrodata, int length)
 {
     eudaq::RawDataEvent ev("TimepixEvent",GetRunNumber(), GetIncreaseEventNumber() );
